requirements.txt
git+https://github.com/Pycord-Development/pycord
aiosqlite
black
jishaku
mystbin.py